refactor: context 훅 분리(#43)
