+++ src/shared/ui/dropdown/dropdown-menu-ltem.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { useDropdownContext } from './dropdown-provider'
+import { useDropdownContext } from './hooks/useDropDownContext'
 /**
  * @description 드롭다운 메뉴 리스트
  *
