+++ src/shared/ui/dropdown/dropdown-menu.tsx
@@ -1,5 +1,5 @@
-import { useDropdownContext } from './dropdown-provider'
 import { cva, type VariantProps } from 'class-variance-authority'
+import { useDropdownContext } from './hooks/useDropDownContext'
 /**
  * @description 드롭다운 메뉴박스
  *
